[WFLY-2609] Consistently use DeploymentManagementResourceRegistration where subsystems may mutate the deployment resource registration tree
